fix : validation of owner of image

+++ backend/rentfriend/src/main/java/org/rentfriend/service/ImageServiceImpl.java
@@ -4,6 +4,7 @@ import jakarta.annotation.PostConstruct;
 import net.coobird.thumbnailator.Thumbnails;
 import net.coobird.thumbnailator.util.ThumbnailatorUtils;
 import org.rentfriend.entity.ImageDB;
+import org.rentfriend.entity.Profile;
 import org.rentfriend.exception.ImageException;
 import org.rentfriend.repository.ImageRepository;
 import org.rentfriend.repository.ProfileRepository;
@@ -25,6 +26,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.security.Principal;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Stream;
 
@@ -34,8 +36,9 @@ public class ImageServiceImpl implements ImageService {
   final Path baseLocation;
   final ImageRepository imageRepository;
   final ProfileRepository profileRepository;
+
   @Autowired
-  public ImageServiceImpl(String location,ImageRepository imageRepository,
+  public ImageServiceImpl(String location, ImageRepository imageRepository,
                           ProfileRepository profileRepository) {
     this.baseLocation = Path.of(location);
     this.imageRepository = imageRepository;
@@ -47,11 +50,11 @@ public class ImageServiceImpl implements ImageService {
   public void init() {
     try {
       Files.createDirectories(baseLocation);
-    }
-    catch (IOException e) {
+    } catch (IOException e) {
       throw new ImageException("Could not initialize storage", e);
     }
   }
+
   @Transactional
   @Override
   public UUID store(MultipartFile file, Principal principal) {
@@ -62,13 +65,13 @@ public class ImageServiceImpl implements ImageService {
       UUID uuid = UUID.randomUUID();
 
       Path destinationFile = this.baseLocation.resolve(
-              Path.of(uuid.toString() +"."+ "JPEG"))
+              Path.of(uuid.toString() + "." + "JPEG"))
           .normalize().toAbsolutePath();
       var profile = profileRepository.findProfileByUser_Username(principal.getName());
-      if(profile.get().getProfileImage()!=null){
+      if (profile.get().getProfileImage() != null) {
         throw new ImageException("File already exists.");
       }
-      imageRepository.save(new ImageDB(uuid,destinationFile.toString(),profile.get()));
+      imageRepository.save(new ImageDB(uuid, destinationFile.toString(), profile.get()));
 
       System.out.println(destinationFile.toString());
       if (!destinationFile.getParent().equals(this.baseLocation.toAbsolutePath())) {
@@ -76,21 +79,20 @@ public class ImageServiceImpl implements ImageService {
         throw new ImageException(
             "Cannot store file outside current directory.");
       }
-      System.out.println("formatss------"+ThumbnailatorUtils.getSupportedOutputFormats());
+      System.out.println("formatss------" + ThumbnailatorUtils.getSupportedOutputFormats());
       try (InputStream inputStream = file.getInputStream()) {
         Thumbnails.of(inputStream)
             .allowOverwrite(true)
-                .scale(0.7)
-                    .outputQuality(0.3)
+            .scale(0.7)
+            .outputQuality(0.3)
             .outputFormat("JPEG")
-                        .toFile(destinationFile.toFile());
+            .toFile(destinationFile.toFile());
 //        Files.copy(inputStream, destinationFile,
 //
 //            StandardCopyOption.REPLACE_EXISTING);
       }
-      return  uuid;
-    }
-    catch (IOException e) {
+      return uuid;
+    } catch (IOException e) {
       throw new ImageException("Failed to store file.", e);
     }
   }
@@ -99,6 +101,7 @@ public class ImageServiceImpl implements ImageService {
   public Stream<Path> loadAll() {
     return Stream.empty();
   }
+
   @Override
   public Path load(String filename) {
     return baseLocation.resolve(filename);
@@ -109,22 +112,20 @@ public class ImageServiceImpl implements ImageService {
   public Resource loadAsResource(String filename) {
     try {
       var img = imageRepository.findById(UUID.fromString(filename));
-      if(img.isEmpty()){
+      if (img.isEmpty()) {
         throw new ImageException("Image not found for: " + filename);
       }
       Path file = Path.of(img.get().getUrl());
-          Resource resource = new UrlResource(file.toUri());
+      Resource resource = new UrlResource(file.toUri());
 
       if (resource.exists() || resource.isReadable()) {
         return resource;
-      }
-      else {
+      } else {
         throw new ImageException(
             "Could not read file: " + filename);
 
       }
-    }
-    catch (MalformedURLException e) {
+    } catch (MalformedURLException e) {
       throw new ImageException("Could not read file: " + filename, e);
     }
   }
@@ -133,12 +134,23 @@ public class ImageServiceImpl implements ImageService {
   public void deleteAll() {
 
   }
-  public void deleteResource(String filename,Principal principal) {
-    var img  = imageRepository.findById(UUID.fromString(filename));
-    if(img.isEmpty()){
+
+  public void deleteResource(String filename, Principal principal) {
+    var img = imageRepository.findById(UUID.fromString(filename));
+    if (img.isEmpty()) {
       throw new ImageException("Image not found for: " + filename);
     }
-    imageRepository.delete(img.get());
+    Optional<Profile> profileO = profileRepository.findProfileByUser_Username(principal.getName());
+    if (profileO.isPresent()) {
+      var profile = profileO.get();
+      if (profile.getProfileImage() != null) {
+        if (profile.getProfileImage().getId().equals(UUID.fromString(filename))) {
+
+          imageRepository.delete(img.get());
+        }
+      }
+    }
+    throw new ImageException("Image not found ");
 
   }
 }
